[refactor] 폴더 구조 재정비 (#193)
* chore : 도메인 중심으로 폴더 구조 변경

* chore : MiniGameReady 폴더 생성

* refactor : 룰렛 돌리기 경로 이름은 roulette/play 로 변경

* chore : feature 안 폴더를 페이지 별로 분리

* refactor: composition 내부의 컴포넌트를 features로 이동

* refactor: feature > ui > modal 컴포넌트를 common 폴더로 이동

* refactor: 경로 에러 수정

* refactor: room 관련 폴더 구조 변경

---------
