+++ frontend/src/features/entry/pages/EntryMenuPage/EntryMenuPage.styled.ts
