+++ frontend/src/features/entry/pages/EntryMenuPage/EntryMenuPage.tsx
