+++ frontend/src/features/entry/pages/EntryNamePage/EntryNamePage.styled.ts
