+++ frontend/src/features/entry/pages/EntryNamePage/EntryNamePage.tsx
