+++ frontend/src/features/home/components/EnterRoomModal/EnterRoomModal.styled.ts
