+++ frontend/src/features/home/components/Splash/Splash.styled.ts
