+++ frontend/src/features/home/components/Splash/Splash.tsx
