+++ frontend/src/features/home/pages/HomePage.tsx
@@ -1,12 +1,12 @@
 import LogoMainIcon from '@/assets/logo-main.svg';
 import Headline3 from '@/components/@common/Headline3/Headline3';
 import RoomActionButton from '@/components/@common/RoomActionButton/RoomActionButton';
-import useModal from '@/features/ui/Modal/useModal';
+import useModal from '@/components/@common/Modal/useModal';
 import Layout from '@/layouts/Layout';
-import Splash from '@/pages/HomePage/Splash/Splash';
+import Splash from '../components/Splash/Splash';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import EnterRoomModal from './EnterRoomModal/EnterRoomModal';
+import EnterRoomModal from '../components/EnterRoomModal/EnterRoomModal';
 import * as S from './HomePage.styled';
 
 const HomePage = () => {
