+++ frontend/src/features/miniGame/components/MiniGameTransition/MiniGameTransition.stories.tsx
