+++ frontend/src/features/miniGame/components/MiniGameTransition/MiniGameTransition.styled.ts
