+++ frontend/src/features/miniGame/components/MiniGameTransition/MiniGameTransition.tsx
