+++ frontend/src/features/miniGame/pages/MiniGameReady/MiniGameReadyPage.styled.ts
