+++ frontend/src/features/miniGame/pages/MiniGameReady/MiniGameReadyPage.tsx
