+++ frontend/src/features/miniGame/pages/MiniGameResultPage/MiniGameResultPage.styled.ts
