+++ frontend/src/features/room/lobby/components/MiniGameSection/MiniGameSection.styled.ts
