+++ frontend/src/features/room/lobby/components/RouletteSection/RouletteSection.tsx
@@ -3,8 +3,8 @@ import StatisticsIcon from '@/assets/statistics-icon.svg';
 import IconButton from '@/components/@common/IconButton/IconButton';
 import SectionTitle from '@/components/@composition/SectionTitle/SectionTitle';
 import { useState } from 'react';
-import ProbabilityList from '../../../components/@composition/ProbabilityList/ProbabilityList';
-import RouletteWheel from '../../../components/@composition/RouletteWheel/RouletteWheel';
+import ProbabilityList from '@/components/@composition/ProbabilityList/ProbabilityList';
+import RouletteWheel from '@/features/roulette/components/RouletteWheel/RouletteWheel';
 import * as S from './RouletteSection.styled';
 
 type RouletteView = 'roulette' | 'statistics';
