+++ frontend/src/features/room/roulette/pages/RoulettePlayPage/RoulettePlayPage.tsx
@@ -8,8 +8,8 @@ import Layout from '@/layouts/Layout';
 import { RouletteView } from '@/types/roulette';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import * as S from './RoulettePage.styled';
-import RoulettePlaySection from './RoulettePlaySection/RoulettePlaySection';
+import * as S from './RoulettePlayPage.styled';
+import RoulettePlaySection from '../../components/RoulettePlaySection/RoulettePlaySection';
 
 const RoulettePage = () => {
   const navigate = useNavigate();
