+++ frontend/src/pages/index.ts
@@ -1,11 +1,11 @@
-export { default as EntryMenuPage } from './EntryMenuPage/EntryMenuPage';
-export { default as EntryNamePage } from './EntryNamePage/EntryNamePage';
-export { default as HomePage } from './HomePage/HomePage';
-export { default as LobbyPage } from './LobbyPage/LobbyPage';
-export { default as MiniGamePlayPage } from './MiniGamePlayPage/MiniGamePlayPage';
-export { default as MiniGameReadyPage } from './MiniGameReadyPage/MiniGameReadyPage';
-export { default as MiniGameResultPage } from './MiniGameResultPage/MiniGameResultPage';
-export { default as NotFoundPage } from './NotFoundPage/NotFoundPage';
-export { default as OrderPage } from './OrderPage/OrderPage';
-export { default as RoulettePage } from './RoulettePage/RoulettePage';
-export { default as RouletteResultPage } from './RouletteResultPage/RouletteResultPage';
+export { default as EntryMenuPage } from '../features/entry/pages/EntryMenuPage/EntryMenuPage';
+export { default as EntryNamePage } from '../features/entry/pages/EntryNamePage/EntryNamePage';
+export { default as HomePage } from '../features/home/pages/HomePage';
+export { default as MiniGamePlayPage } from '../features/miniGame/pages/MiniGamePlayPage/MiniGamePlayPage';
+export { default as MiniGameReadyPage } from '../features/miniGame/pages/MiniGameReady/MiniGameReadyPage';
+export { default as MiniGameResultPage } from '../features/miniGame/pages/MiniGameResultPage/MiniGameResultPage';
+export { default as NotFoundPage } from '../features/notFound/pages/NotFoundPage';
+export { default as LobbyPage } from '../features/room/lobby/pages/LobbyPage';
+export { default as OrderPage } from '../features/room/order/pages/OrderPage';
+export { default as RoulettePage } from '../features/room/roulette/pages/RoulettePlayPage/RoulettePlayPage';
+export { default as RouletteResultPage } from '../features/room/roulette/pages/RouletteResultPage/RouletteResultPage';
